server folder is structured
