+++ server/index.js
@@ -1,10 +1,10 @@
-const connectDB = require('./db/mongoose.js');
+const connectDB = require('./db/connect');
 const express = require('express');
 const app = express();
 const port = process.env.PORT || 4000;
-const userRoute = require('./routes/user');
-const doctorRoute = require('./routes/doctor');
-const resetRoute = require('./routes/reset.js');
+const userRoute = require('./api/routes/user');
+const doctorRoute = require('./api/routes/doctor');
+const resetRoute = require('./api/routes/reset');
 const dotenv = require('dotenv');
 const cors = require('cors');
